Fix stack caller for error reporting
Rather than using a hardcoded stack skip index, this change walks up the stack
to find the first caller in a package outside of a skip list.

By default, only logrus is skipped. However, this can be overridden with
stackdriver.WithStackSkip("package").

This is useful for skipping packages that might be wrapping logrus. It
also supports vendoring.

It's very similar to #3 except it doesn't rely on regexp.

Updated the tests which had incorrect line number and functions.